chore: Update useRefreshToken.hook.tsx to use 'access_token' instead of 'token'
Refactor useRefreshToken.hook.tsx to update the assignment of the 'access_token' property in the session.user object. This change aligns with the recent updates to the next-auth.d.ts file, which renamed the 'token' property to 'access_token' and added new properties like 'refresh_token' and 'lf_id' to the user object.